Fix for sign mismatch warning/error
git-svn-id: http://skia.googlecode.com/svn/trunk@7628 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ debugger/SkDebugCanvas.cpp
@@ -109,7 +109,7 @@ SkPMColor OverdrawXferModeProc(SkPMColor src, SkPMColor dst) {
         SkPackARGB32(0xFF, 255,  0, 0)
     };
 
-    for (int i = 0; i < SK_ARRAY_COUNT(gTable)-1; ++i) {
+    for (size_t i = 0; i < SK_ARRAY_COUNT(gTable)-1; ++i) {
         if (gTable[i] == dst) {
             return gTable[i+1];
         }
